chore: introduced bitcoin module structure, #6250
