+++ stacks-node/src/tests/epoch_22.rs
@@ -19,10 +19,10 @@ use stacks_common::util::secp256k1::Secp256k1PublicKey;
 use stacks_common::util::sleep_ms;
 
 use super::neon_integrations::get_account;
+use crate::burnchains::bitcoin::core_controller::BitcoinCoreController;
 use crate::neon_node::StacksNode;
 use crate::stacks_common::types::Address;
 use crate::stacks_common::util::hash::bytes_to_hex;
-use crate::tests::bitcoin_regtest::BitcoinCoreController;
 use crate::tests::epoch_21::wait_pox_stragglers;
 use crate::tests::neon_integrations::*;
 use crate::tests::*;
