+++ stacks-node/src/tests/epoch_23.rs
@@ -24,7 +24,7 @@ use stacks::core::test_util::make_contract_call;
 use stacks::core::{self, EpochList, STACKS_EPOCH_MAX};
 use stacks_common::util::sleep_ms;
 
-use crate::tests::bitcoin_regtest::BitcoinCoreController;
+use crate::burnchains::bitcoin::core_controller::BitcoinCoreController;
 use crate::tests::neon_integrations::*;
 use crate::tests::*;
 use crate::{neon, BitcoinRegtestController, BurnchainController};
