+++ stacks-node/src/tests/neon_integrations.rs
@@ -84,8 +84,8 @@ use stacks_common::util::{get_epoch_time_ms, get_epoch_time_secs, sleep_ms};
 use tokio::io::{AsyncReadExt as _, AsyncWriteExt as _};
 use tokio::net::{TcpListener, TcpStream};
 
-use super::bitcoin_regtest::BitcoinCoreController;
 use super::{ADDR_4, SK_1, SK_2, SK_3};
+use crate::burnchains::bitcoin::core_controller::BitcoinCoreController;
 use crate::burnchains::bitcoin_regtest_controller::{self, addr2str, BitcoinRPCRequest, UTXO};
 use crate::neon_node::RelayerThread;
 use crate::operations::BurnchainOpSigner;
